+++ lib/waterdrop/errors.rb
@@ -60,4 +60,8 @@ module WaterDrop
       end
     end
   end
+
+  # Alias so we can have a nicer API to abort transactions
+  # This makes referencing easier
+  AbortTransaction = Errors::AbortTransaction
 end