+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # WaterDrop changelog
 
 ## 2.7.4 (Unreleased)
-- [Maintenance] Alias `WaterDrop::Errors::AbortTransaction` with `WaterDrop::Rollback` and `WaterDrop::AbortTransaction`.
+- [Maintenance] Alias `WaterDrop::Errors::AbortTransaction` with `WaterDrop::AbortTransaction`.
 - [Maintenance] Lower the precision reporting to 100 microseconds in the logger listener.
 - [Fix] Consumer consuming error: Local: Erroneous state (state) post break flow in transaction.
 - [Change] Require 'karafka-core' `>= 2.4.3`